+++ src/index.tsx
@@ -3,18 +3,14 @@ React-Quill
 https://github.com/zenoamaro/react-quill
 */
 
-import React from 'react';
-import ReactDOM from 'react-dom';
-import isEqual from 'lodash/isEqual';
+import React, { createRef } from 'react';
+import { isEqual } from 'lodash-es';
 
-import Quill, {
-  QuillOptionsStatic,
-  DeltaStatic,
-  RangeStatic,
-  BoundsStatic,
-  StringMap,
-  Sources,
-} from 'quill';
+import Quill, { type EmitterSource, type Range as RangeStatic, QuillOptions as QuillOptionsStatic } from 'quill';
+import type DeltaStatic from 'quill-delta';
+
+export { Quill }
+export type { DeltaStatic, EmitterSource, RangeStatic, QuillOptionsStatic }
 
 // Merged namespace hack to export types along with default object
 // See: https://github.com/Microsoft/TypeScript/issues/2719
@@ -33,26 +29,26 @@ namespace ReactQuill {
     defaultValue?: Value,
     formats?: string[],
     id?: string,
-    modules?: StringMap,
+    modules?: QuillOptions['modules'],
     onChange?(
       value: string,
       delta: DeltaStatic,
-      source: Sources,
+      source: EmitterSource,
       editor: UnprivilegedEditor,
     ): void,
     onChangeSelection?(
       selection: Range,
-      source: Sources,
+      source: EmitterSource,
       editor: UnprivilegedEditor,
     ): void,
     onFocus?(
       selection: Range,
-      source: Sources,
+      source: EmitterSource,
       editor: UnprivilegedEditor,
     ): void,
     onBlur?(
       previousSelection: Range,
-      source: Sources,
+      source: EmitterSource,
       editor: UnprivilegedEditor,
     ): void,
     onKeyDown?: React.EventHandler<any>,
@@ -61,7 +57,6 @@ namespace ReactQuill {
     placeholder?: string,
     preserveWhitespace?: boolean,
     readOnly?: boolean,
-    scrollingContainer?: string | HTMLElement,
     style?: React.CSSProperties,
     tabIndex?: number,
     theme?: string,
@@ -69,12 +64,13 @@ namespace ReactQuill {
   }
 
   export interface UnprivilegedEditor {
-    getLength(): number;
-    getText(index?: number, length?: number): string;
-    getHTML(): string;
-    getBounds(index: number, length?: number): BoundsStatic;
-    getSelection(focus?: boolean): RangeStatic;
-    getContents(index?: number, length?: number): DeltaStatic;
+    getLength: Quill['getLength'];
+    getText: Quill['getText'];
+    getHTML: () => string;
+    getSemanticHTML: Quill['getSemanticHTML'];
+    getBounds: Quill['getBounds'];
+    getSelection: Quill['getSelection'];
+    getContents: Quill['getContents'];
   }
 }
 
@@ -90,6 +86,7 @@ interface ReactQuillState {
 }
 
 class ReactQuill extends React.Component<ReactQuillProps, ReactQuillState> {
+  editingAreaRef = createRef<any>();
 
   static displayName = 'React Quill'
 
@@ -143,11 +140,6 @@ class ReactQuill extends React.Component<ReactQuillProps, ReactQuillState> {
   The Quill Editor instance.
   */
   editor?: Quill
-
-  /*
-  Reference to the element holding the Quill editing area.
-  */
-  editingArea?: React.ReactInstance | null
 
   /*
   Tracks the internal value of the Quill editor
@@ -313,7 +305,6 @@ class ReactQuill extends React.Component<ReactQuillProps, ReactQuillState> {
       modules: this.props.modules,
       placeholder: this.props.placeholder,
       readOnly: this.props.readOnly,
-      scrollingContainer: this.props.scrollingContainer,
       tabIndex: this.props.tabIndex,
       theme: this.props.theme,
     };
@@ -328,7 +319,7 @@ class ReactQuill extends React.Component<ReactQuillProps, ReactQuillState> {
   Creates an editor on the given element. The editor will be passed the
   configuration, have its events bound,
   */
-  createEditor(element: Element, config: QuillOptions) {
+  createEditor(element: HTMLElement, config: QuillOptions) {
     const editor = new Quill(element, config);
     if (config.tabIndex != null) {
       this.setEditorTabIndex(editor, config.tabIndex);
@@ -384,7 +375,7 @@ class ReactQuill extends React.Component<ReactQuillProps, ReactQuillState> {
     this.value = value;
     const sel = this.getEditorSelection();
     if (typeof value === 'string') {
-      editor.setContents(editor.clipboard.convert(value));
+      editor.setContents(editor.clipboard.convert({html: value}));
     } else {
       editor.setContents(value);
     }
@@ -423,27 +414,25 @@ class ReactQuill extends React.Component<ReactQuillProps, ReactQuillState> {
   makeUnprivilegedEditor(editor: Quill) {
     const e = editor;
     return {
-      getHTML:      () => e.root.innerHTML,
-      getLength:    e.getLength.bind(e),
-      getText:      e.getText.bind(e),
-      getContents:  e.getContents.bind(e),
-      getSelection: e.getSelection.bind(e),
-      getBounds:    e.getBounds.bind(e),
+      getHTML:         () => e.root.innerHTML,
+      getSemanticHTML: e.getSemanticHTML.bind(e),
+      getLength:       e.getLength.bind(e),
+      getText:         e.getText.bind(e),
+      getContents:     e.getContents.bind(e),
+      getSelection:    e.getSelection.bind(e),
+      getBounds:       e.getBounds.bind(e),
     };
   }
 
-  getEditingArea(): Element {
-    if (!this.editingArea) {
-      throw new Error('Instantiating on missing editing area');
-    }
-    const element = ReactDOM.findDOMNode(this.editingArea);
+  getEditingArea(): HTMLElement {
+    const element = this.editingAreaRef.current;
     if (!element) {
       throw new Error('Cannot find element for editing area');
     }
     if (element.nodeType === 3) {
       throw new Error('Editing area cannot be a text node');
     }
-    return element as Element;
+    return element as HTMLElement;
   }
 
   /*
@@ -455,9 +444,7 @@ class ReactQuill extends React.Component<ReactQuillProps, ReactQuillState> {
 
     const properties = {
       key: generation,
-      ref: (instance: React.ReactInstance | null) => {
-        this.editingArea = instance
-      },
+      ref: this.editingAreaRef,
     };
 
     if (React.Children.count(children)) {
@@ -492,7 +479,7 @@ class ReactQuill extends React.Component<ReactQuillProps, ReactQuillState> {
     eventName: 'text-change' | 'selection-change',
     rangeOrDelta: Range | DeltaStatic,
     oldRangeOrDelta: Range | DeltaStatic,
-    source: Sources,
+    source: EmitterSource,
   ) => {
     if (eventName === 'text-change') {
       this.onEditorChangeText?.(
@@ -513,7 +500,7 @@ class ReactQuill extends React.Component<ReactQuillProps, ReactQuillState> {
   onEditorChangeText(
     value: string,
     delta: DeltaStatic,
-    source: Sources,
+    source: EmitterSource,
     editor: UnprivilegedEditor,
   ): void {
     if (!this.editor) return;
@@ -536,7 +523,7 @@ class ReactQuill extends React.Component<ReactQuillProps, ReactQuillState> {
 
   onEditorChangeSelection(
     nextSelection: RangeStatic,
-    source: Sources,
+    source: EmitterSource,
     editor: UnprivilegedEditor,
   ): void {
     if (!this.editor) return;
@@ -575,6 +562,4 @@ function postpone(fn: (value: void) => void) {
   Promise.resolve().then(fn);
 }
 
-// Compatibility Export to avoid `require(...).default` on CommonJS.
-// See: https://github.com/Microsoft/TypeScript/issues/2719
-export = ReactQuill;
+export default ReactQuill;